Move `getLegendaryVersion`, `getGogdlVersion`, `getNileVersion` and `getHericVersion` outside of utils.ts
The behavior of everything but `getHeroicVersion` stayed the same.
Heroic's own version codenames will now live inside `package.json` (right next
to the version number) for convenience
