+++ package.json
@@ -1,6 +1,10 @@
 {
   "name": "heroic",
   "version": "2.9.1",
+  "versionNames": {
+    "stable": "Boa Hancock",
+    "beta": "Caesar Clown"
+  },
   "private": true,
   "main": "build/electron/main.js",
   "homepage": "./",
